strutil: move out ColoredString to colored_string.py
